pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "FastAPI Production Boilerplate"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
dependencies = [
    # Core
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "gunicorn>=21.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.11.0",
    "asyncpg>=0.28.0",  # PostgreSQL async driver
    "psycopg2-binary>=2.9.7",  # For alembic to work
    "oracledb>=1.3.0",  # Oracle async driver
    "pgvector>=0.2.0",  # pgvector Python client
    
    # Caching & Rate Limiting
    "redis>=4.6.0",
    "fastapi-cache2>=0.2.1",
    "fastapi-limiter>=0.1.5",
    
    # Storage
    "miniopy-async>=1.16",
    
    # Logging & Monitoring
    "loguru>=0.7.0",
    
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.24.1",
    
    # Utilities
    "tenacity>=8.2.2",  # For retrying operations
    "pydantic-extra-types>=2.0.0",  # Additional Pydantic types
]

[project.optional-dependencies]
dev = [
    # Development tools
    "black>=23.3.0",
    "ruff>=0.0.275",
    "mypy>=1.4.1",
    "pre-commit>=3.3.3",
    "pytest-cov>=4.1.0",
    "faker>=18.13.0",
    "locust>=2.15.1",
]

[tool.uv]
package-manager = "uv"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "B", "I", "N", "UP", "YTT", "ANN", "C4", "ICN", "PIE", "PYI"]
ignore = ["ANN401", "ANN101"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"alembic/env.py" = ["ANN"]
"tests/*" = ["ANN"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "sqlalchemy.*",
    "alembic.*",
    "fastapi_cache2.*",
    "fastapi_limiter.*",
    "miniopy_async.*",
    "jose.*",
    "passlib.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"